login functionality completed

+++ home/views.py
@@ -1,8 +1,12 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
 from django.http import HttpResponse
 from django.db.models import Q
 from .models import Hive, Topic
 from .forms import HiveForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 
 '''
 hive = room
@@ -16,21 +20,36 @@ buzz = name
 
 # Create your views here.
 def loginView(request):
+  username, password = '', ''
+  
+  page = 'login'
+  if request.user.is_authenticated:
+    return redirect('homepage')
+   
   if request.method == 'POST':
     username = request.POST.get('username')
     password = request.POST.get('password')
     
-    try:
-      user = user.objects.get(username=username)
-      
-    except:
-      pass
-      
-    
-  context={'username': username, 'password': password}
+    user = authenticate(request, username=username, password=password)
+    if user is not None:
+      login(request, user)
+      return redirect('homepage')
+
+    else:
+      messages.error(request, "We could not find your username")
+
+  context={'username': username, 'password': password, 'page': login}
   return render(request, 'home/logreg.html', context)
   
+def logoutView(request):
+  
+  logout(request)
+  return redirect('homepage')
 
+def registerUser(request):
+  context={'page': 'register'}
+  return render(request, 'home/logreg.html', context )
+  
 
 def home(request):
   ''' 
@@ -60,6 +79,8 @@ def hive(request, pk):
   return render(request, 'home/hive.html', context)
 
 # CRUD Operations
+
+@login_required(login_url='login')
 def createHive(request):
   form = HiveForm()
   if request.method == 'POST':
@@ -71,10 +92,13 @@ def createHive(request):
   context = {"form": form}
   return render(request, 'home/hiveForm.html', context)
 
-
+@login_required(login_url='login')
 def updateHive(request, pk):
   hive = Hive.objects.get(id=pk)
   form = HiveForm(instance=hive) #pre-fill with values
+  
+  if request.user != hive.creator:
+    return HttpResponse("Nah fam i can't allow it")
   
   if request.method == 'POST':  #ensure the current editable hive is updated
     form = HiveForm(request.POST, instance=hive)
@@ -84,8 +108,13 @@ def updateHive(request, pk):
     
   return render(request, 'home/hiveForm.html', {'form': form})
 
+@login_required(login_url='login')
 def deleteHive(request, pk):
   hive = Hive.objects.get(id=pk)
+  
+  if request.user != hive.creator:
+    return HttpResponse("Nah fam i can't allow it")
+  
   if request.method == "POST":
     hive.delete()
     return redirect('homepage')
